pyproject.toml
[project]
name = "fasthtml-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
